Add files via upload
Airbnb, Inc. is an American company operating an online marketplace for homestays. The attached report is a pursuit to design and deliver an end-to-end business intelligence solution for them.
The attached report provides a way to track Prices for different Stay locations, available Bedroom capacity, and Revenue details to AirBnb in an easy-to-understand format.